ynputconnect endpoints authentication

+++ api/connect/__init__.py
@@ -1,6 +1,8 @@
 import httpx
 from fastapi import APIRouter
 
+from ayon_server.api.dependencies import CurrentUser
+from ayon_server.exceptions import BadRequestException, ForbiddenException
 from ayon_server.lib.postgres import Postgres
 from ayon_server.types import Field, OPModel
 
@@ -11,7 +13,7 @@ router = APIRouter(
 
 
 @router.get("")
-async def get_ynput_connect_info():
+async def get_ynput_connect_info(user: CurrentUser):
     res = await Postgres.fetch(
         """
         SELECT value FROM secrets
@@ -20,10 +22,11 @@ async def get_ynput_connect_info():
     )
 
     if not res:
-        raise Exception("Ynput connect key not found")
+        raise BadRequestException("Ynput connect key not found")
 
     key = res[0]["value"]
 
+    # TODO: handle errors
     async with httpx.AsyncClient() as client:
         res = await client.get(f"https://auth.ayon.cloud/info?key={key}")
 
@@ -35,7 +38,10 @@ class YnputConnectRequestModel(OPModel):
 
 
 @router.post("")
-async def set_ynput_connect_key(request: YnputConnectRequestModel):
+async def set_ynput_connect_key(request: YnputConnectRequestModel, user: CurrentUser):
+    if not user.is_admin:
+        raise ForbiddenException("Only admins can set the Ynput connect key")
+
     await Postgres.execute(
         """
         INSERT INTO secrets (name, value)
@@ -49,7 +55,9 @@ async def set_ynput_connect_key(request: YnputConnectRequestModel):
 
 
 @router.delete("")
-async def delete_ynput_connect_key():
+async def delete_ynput_connect_key(user: CurrentUser):
+    if not user.is_admin:
+        raise ForbiddenException("Only admins can set the Ynput connect key")
     await Postgres.execute(
         """
         DELETE FROM secrets
